STIG-PGD/requirements.txt
# STIG-PGD Requirements
# Deep Learning
torch>=1.12.0
torchvision>=0.13.0
torchsummary>=1.5.1

# Image Processing
Pillow>=9.0.0
opencv-python>=4.5.0

# Scientific Computing
numpy>=1.21.0
scipy>=1.7.0

# Visualization
matplotlib>=3.5.0

# Utilities
tqdm>=4.62.0

# Metrics
pytorch-fid>=0.3.0

# GUI
gradio>=4.0.0

